[♻️refactor]: 폴더명 대문자로 통일, Info 폴더 네이밍을 Information으로 변경
