+++ src/components/Requirement/RequirementSection.tsx
@@ -8,7 +8,7 @@ import { theme } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
 
 import ResizeHandle from '../Shared/ResizeHandle';
-import GPTPrompt from '../gpt/GptPrompt';
+import GPTPrompt from '../Gpt/GptPrompt';
 import RequirementList from './RequirementList';
 
 type RequirementSectionProps = {
