+++ src/components/gpt/DisconnectedPrompt.tsx
@@ -4,8 +4,8 @@ import { theme, ThemeType } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
 
 import CodeBlock from './CodeBlock';
-import GPTAPIKeyInfo from './info/GPTAPIKeyInfo';
-import GPTSearchInfo from './info/GPTSearchInfo';
+import GPTAPIKeyInfo from './Information/GPTAPIKeyInfo';
+import GPTSearchInfo from './Information/GPTSearchInfo';
 
 type DisconnectedPromptProps = {
   requirement?: RequirementStateType;
