+++ src/components/gpt/Information/GPTSearchInfo.tsx
